Inline _store.py's store-glue helpers into _factor.py
has_write/with_store had exactly one consuming module (_factor.py); fold them
in directly since the separate file bought no reuse.
